Update area of the circle.c

+++ area of the circle.c
@@ -4,7 +4,7 @@ int main()
 	int r;
 	printf("enter the value of radius:");
 	scanf("%d",&r);
-	int a= 3.14*r*r;
-	printf("the area of circle  is %d",a);
+	float a= 3.14*r*r;
+	printf("the area of circle  is %f",a);
 return 0;	
 }
